Decimal precision scale datatype change (#2532)
* Change decimal array precison and scale to u8

* bug fixes

+++ arrow/benches/cast_kernels.rs
@@ -82,7 +82,7 @@ fn build_utf8_date_time_array(size: usize, with_nulls: bool) -> ArrayRef {
     Arc::new(builder.finish())
 }
 
-fn build_decimal128_array(size: usize, precision: usize, scale: usize) -> ArrayRef {
+fn build_decimal128_array(size: usize, precision: u8, scale: u8) -> ArrayRef {
     let mut rng = seedable_rng();
     let mut builder = Decimal128Builder::with_capacity(size, precision, scale);
 
@@ -92,7 +92,7 @@ fn build_decimal128_array(size: usize, precision: usize, scale: usize) -> ArrayR
     Arc::new(builder.finish())
 }
 
-fn build_decimal256_array(size: usize, precision: usize, scale: usize) -> ArrayRef {
+fn build_decimal256_array(size: usize, precision: u8, scale: u8) -> ArrayRef {
     let mut rng = seedable_rng();
     let mut builder = Decimal256Builder::with_capacity(size, precision, scale);
     let mut bytes = [0; 32];
